+++ CHANGELOG.md
@@ -1,10 +1,15 @@
-*** 0.1.1
+\*\*\* 0.1.1
 
 - Changed default filter bindings changed to make projectile not a hard dependency
 - Added `frog-jump-buffer-include-current-buffer` option
 - Allow for overriding `posframe` related options
 
-*** 0.1.2
+\*\*\* 0.1.2
 
 - Add support for filtering by `recentf` and opening closed buffers
 - Change default filter bindings again to make room for `recentf`
+
+\*\*\* 0.1.3
+
+- Add similarly named buffers filters
+- Always show current project filter and error on selection if `projectile` is not installed